Set Up Container Div For Main App
In overcast's repo, we set up our "frame" around the entire app using
the `App` class. Since we'll now be having a separate landing page for
non-logged in users, create a `MainApp` class so that our styling
doesn't interfere with the landing page.

Also, use `id`s for the `App` and `MainApp`, because they should be
unique.