Landing Page: Fix Background Height Problem
The content was overflowing past the `Background` container. It turned
out that the height metric in `index.css` was unnecessary and preventing
the background div from expanding. Remove that. Also, make the
background div a minimum of the window's height so there's no white
space when the user is zoomed out.