Add Method for Rendering Notices
Add a method for rendering notices about the user data popup that we
want to show the user. Currently, this method only displays one notice
to the user in the event they don't have any apps with which to share
data.

Use this method in `renderDataPopup`. Now, instead of simply rendering
the chart as it did before, it renders the chart AND any notices that
need to be displayed.

This is step two of the Wrap Method pattern applied.